websocket: Extract websocket parser to separate file
This will allow to implement Websocket client that reuses existing code.
